backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
numpy
ffmpeg-python
openai-whisper
torch 
fastapi